wip: account balances
Need to decide on account's existence.

What does it mean to register an account? This only seems to make sence in the
context of a client of some persistence service. For the service of calculation,
Only transaction information is needed to query about, **known**, accounts. It
does not make sence that a client wishes to discover **unknown** accounts.