complete oop style api up to what Book has to offer
Came accross the lifetime issue again when `Book::running_balance` returns
`Transaction<'a>` where `'a` is of the book. This too does not cross the
`std::cell::Ref<'_, T>` boundry. However, there is a way to get to the underlying
lifetime through an unstable feature 'cell_leak'

see https://github.com/rust-lang/rust/issues/69099

This allows the borrow checker to life-check!