eliminate `Rc`s and `Weak`s
This commit does the opposite of the previous commit "eliminate lifetimes".
Usage of `Rc`, and then `Weak` pronounces the tasks of validating a 'pointer'
is valid. This is, infact, the job of the borrow checker!

In attempt to take advantage of borrow checking so that obsolete references
are rejected by the compiler. Such that

```rust
#[test]
fn transfer() {
    let mut book = Book::new();
    book.new_account("acc");
    let acc: &Account = book.account_with_name("acc").unwrap();
    let thb = THB();
    book.transfer(acc, acc, thb.of_major(5));
    // the following line would not compile because `acc` was borrowed from book
    // before book was mutably borrowed by the method `transfer`
    assert_eq!(book.balance(acc), Money::none());
}
```
However, this was a failed attempt. Use of `acc` is rejected at the invokation of `transfer`
and compiler rejects usage of it in 'the call' of `transfer`. Because the call _mutably_
borrows from `book`.

Further modifications to `Book::tranfer` so that it takes `from_or_to_account_name: &str`s
instead of `&Account`s only reveals another issue: feilds of a struct can not borrow
from the same struct!

A contrived demonstration of this issue is at https://github.com/gomain/cannot-borrow-as-immutable

This commit is to keep record of the context by which `Transaction` can no longer keep
references to accounts if transactions are to be owned by the book and also are accounts.