refer to accounts by id
This commit solves the _self aliasing_ issue where attempt to use a borrowed value from self as
argumets to methods that mutably borrow self. as in

```rust
    let mut book = Book::new();
    let borrowed = book.borrow();
    book.mutate(borrowed); // this will not compile
```
This was the case for accounts. For this `Account` now have an `id` field. Accounts in the book are
referred to by id, and looked up in it's internal collection of accounts.

This solution implies possibility of failure in the calls to the api of the book; a provided id, may
not exist. Thus all failable apis return some `Result<_, E>` where E is a custom error type.

It is noticed that referring to accounts by ids are only necessary for the book api. Not for the
`accounting` module. `Transaction`s could have continued to keep references of `Account`'s. Only in
`Book`, account ids are introduced.