decouple `AccountId` from accounting
Having realized that accounts needing tobe referred to by ids, instead of by reference - for the book
api - was only a concern imposed by the book. This shouldn't have affected the implemetation of the
accounting module. It was only the book that could not let a refference to one of its accounts then
later accept it for making transactions. In the accounting layer, this was not a problem! So to keep
the accounting layer as pure as possible, this commit reverts `Transaction` to keep referrences of its
`from` and `to` accounts. And `Account::balance` takes a slice of such transactions.

Associating accounts with ids is the bussiness of the book. For users of the book, transactions are from
& to account ids. Transformation from `book::Transaction` to `accounting::Transaction` is trivial.

However, ironicaly, since an account now has no data. An atomic internal-id generator is needed so that accounts
can be compared for equality by deriving `Eq`.