generalize running balance calculations in module `accounting`
This is to suppport the usecase of clients of the `accounting` module representing transactions with
a different structure than the internal `Transaction` type. Essentially this is to provide a method
returning an iterater over `(&T, Money)` given an iterator over `&T` and a converter function `fn
(&T) -> Transaction` where T is the external representation of transaction.

On top of that, the conversion function may fail. Thus the running balance calculation may
fail. This leads to some excessive use of type parameters in the running balance method. One for the
iterator, one for the external transaction, one for the convertion function, and another for the
error.